backend/basic-mindbot/requirements.txt
livekit-agents[openai, cartesia, elevenlabs, deepgram, silero, turn-detector, mcp]>=1.0
python-dotenv>=1.0
duckduckgo-search>=8.0
aiohttp>=3.9.0